chore: Slightly breaking changes to the API and we're in the 2.0 era
feat: Boot-starter adds support for multiton

performance: Fine tune multiton

test: Some basic test&benchmark to the new feature

refactor: A lot bit

+++ stun4j-guid-core/src/main/java/com/stun4j/guid/core/LocalZkGuid.java
@@ -46,29 +46,29 @@ public class LocalZkGuid {
     LocalGuid.instance().reset(nodeNewInfo);
   };
 
-  public static LocalGuid init(String zkConnectStr, int digits, long datacenterIdBits, long workerIdBits, long seqBits,
+  public static LocalGuid init(String zkConnectStr, int digits, int datacenterIdBits, int workerIdBits, int seqBits,
       boolean fixedDigitsEnabled) throws Exception {
     return init(zkConnectStr, null, digits, datacenterIdBits, workerIdBits, seqBits, fixedDigitsEnabled, null);
   }
 
-  public static LocalGuid init(String zkConnectStr, String zkNamespace, int digits, long datacenterIdBits,
-      long workerIdBits, long seqBits, boolean fixedDigitsEnabled, String ipStartWith) throws Exception {
+  public static LocalGuid init(String zkConnectStr, String zkNamespace, int digits, int datacenterIdBits,
+      int workerIdBits, int seqBits, boolean fixedDigitsEnabled, String ipStartWith) throws Exception {
     Pair<Integer, Integer> nodeInfo = ZkGuidNode.start(zkConnectStr, ON_RECONNECT_FN, zkNamespace, digits,
         datacenterIdBits, workerIdBits, seqBits, fixedDigitsEnabled, ipStartWith);
 
     return doInit(nodeInfo, digits, datacenterIdBits, workerIdBits, seqBits, fixedDigitsEnabled);
   }
 
-  public static LocalGuid init(Builder zkClientBuilder, int digits, long datacenterIdBits, long workerIdBits,
-      long seqBits, boolean fixedDigitsEnabled, String ipStartWith) throws Exception {
+  public static LocalGuid init(Builder zkClientBuilder, int digits, int datacenterIdBits, int workerIdBits, int seqBits,
+      boolean fixedDigitsEnabled, String ipStartWith) throws Exception {
     Pair<Integer, Integer> nodeInfo = ZkGuidNode.start(zkClientBuilder, ON_RECONNECT_FN, digits, datacenterIdBits,
         workerIdBits, seqBits, fixedDigitsEnabled, ipStartWith);
 
     return doInit(nodeInfo, digits, datacenterIdBits, workerIdBits, seqBits, fixedDigitsEnabled);
   }
 
-  private static LocalGuid doInit(Pair<Integer, Integer> nodeInfo, int digits, long datacenterIdBits, long workerIdBits,
-      long seqBits, boolean fixedDigitsEnabled) {
+  private static LocalGuid doInit(Pair<Integer, Integer> nodeInfo, int digits, int datacenterIdBits, int workerIdBits,
+      int seqBits, boolean fixedDigitsEnabled) {
     return LocalGuid.init(nodeInfo.getLeft(), nodeInfo.getRight(), digits, datacenterIdBits, workerIdBits, seqBits,
         fixedDigitsEnabled);
   }
